Now push all that logic into a standalone class
